Add halt() and some error codes
Define a halt function that can be implemented by the test framework
for use on error paths before exit() works. Also add some error codes
that can be passed to halt() in case the serial isn't working either.
Then, on register inspection of the halted guest we may be able to
quickly determine the problem without having to find the halt()
call-site. The error codes may of course also be used with exit().

Signed-off-by: Andrew Jones <drjones@redhat.com>